chore(beam): remove legacy border beam -> modernized
